% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggaligner.R
\name{ggaligner}
\alias{ggaligner}
\title{Visualize the alignment object from msa package}
\usage{
ggaligner(alignment,start=1,end=10,color="Clustal",font="helvetical",label_font = 12)
}
\arguments{
\item{alignment}{alignment object returned from msa package}

\item{start}{start position of the desired alignment region}

\item{end}{end position of the desired alignment region}

\item{color}{color scheme to use ex: Clustal, Chemistry_AA, Shapely_AA}

\item{font}{font family to use ex: helvetical, TimesNewRoman}

\item{label_font}{font size of sequence names}
}
\value{
An enhanced plot for the alignment using ggplot2 and ggmsa packages
}
\description{
Generating a plot for msa object of DNA or Protein sequence
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
