% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seqtochar.R
\name{Seqtochar}
\alias{Seqtochar}
\title{Convert string to a character vector}
\usage{
Seqtochar(x)
}
\arguments{
\item{x}{a sequence of DNA or Protein}
}
\value{
this function is mainly used by ggaligner function to convert the input sequence to character vector
}
\description{
convert string that can be DNA or protein sequence to a character vector
}
\examples{
Seqtochar("ATGACATAAT")

}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
